fixing task 3

+++ functions_print2.c
@@ -9,7 +9,7 @@ int p_bin(va_list list)
 	unsigned long long int binarynum = 0;
 	unsigned long long int i = 1, rem;
 	unsigned long long int temp = 1;
-	unsigned int decimalnum = va_arg(list, unsigned int);
+	unsigned long long int decimalnum = va_arg(list, unsigned long long int);
 
 	while (decimalnum != 0)
 	{
